+++ src/Integration/RequestIntegration.php
@@ -42,17 +42,24 @@ final class RequestIntegration implements IntegrationInterface
     private $options;
 
     /**
+     * @var RequestFetcherInterface PSR-7 request fetcher
+     */
+    private $requestFetcher;
+
+    /**
      * Constructor.
      *
-     * @param Options $options The client options
+     * @param Options|null                 $options        The client options
+     * @param RequestFetcherInterface|null $requestFetcher PSR-7 request fetcher
      */
-    public function __construct(?Options $options = null)
+    public function __construct(?Options $options = null, ?RequestFetcherInterface $requestFetcher = null)
     {
         if (null !== $options) {
             @trigger_error(sprintf('Passing the options as argument of the constructor of the "%s" class is deprecated since version 2.1 and will not work in 3.0.', self::class), E_USER_DEPRECATED);
         }
 
         $this->options = $options;
+        $this->requestFetcher = $requestFetcher ?? new RequestFetcher();
     }
 
     /**
@@ -98,7 +105,7 @@ final class RequestIntegration implements IntegrationInterface
     private function processEvent(Event $event, Options $options, ?ServerRequestInterface $request = null): void
     {
         if (null === $request) {
-            $request = isset($_SERVER['REQUEST_METHOD']) && \PHP_SAPI !== 'cli' ? ServerRequest::fromGlobals() : null;
+            $request = $this->requestFetcher->fetchRequest();
         }
 
         if (null === $request) {